Open/master

+++ georchestra-signalement-service/src/main/java/org/georchestra/signalement/service/sm/impl/InitializationServiceImpl.java
@@ -1,22 +1,22 @@
 /**
- * 
+ *
  */
 package org.georchestra.signalement.service.sm.impl;
 
-import java.io.FileInputStream;
-import java.math.BigDecimal;
-import java.util.ArrayList;
-import java.util.List;
-
 import org.activiti.engine.ProcessEngine;
 import org.activiti.engine.RepositoryService;
+import org.activiti.engine.TaskService;
 import org.activiti.engine.repository.Deployment;
 import org.activiti.engine.repository.ProcessDefinition;
 import org.activiti.engine.repository.ProcessDefinitionQuery;
+import org.activiti.engine.task.Task;
 import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.georchestra.signalement.core.common.DocumentContent;
+import org.georchestra.signalement.core.dto.ContextDescription;
+import org.georchestra.signalement.service.common.ErrorMessageConstants;
 import org.georchestra.signalement.service.exception.InitializationException;
+import org.georchestra.signalement.service.exception.InvalidDataException;
 import org.georchestra.signalement.service.mapper.workflow.ProcessDefinitionMapper;
 import org.georchestra.signalement.service.sm.InitializationService;
 import org.slf4j.Logger;
@@ -24,6 +24,11 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import org.springframework.util.MimeTypeUtils;
+
+import java.io.FileInputStream;
+import java.util.Comparator;
+import java.util.List;
+import java.util.stream.Collectors;
 
 /**
  * @author FNI18300
@@ -36,9 +41,12 @@ public class InitializationServiceImpl implements InitializationService {
 
 	@Autowired
 	private ProcessEngine processEngine;
-	
+
 	@Autowired
 	private ProcessDefinitionMapper processDefinitionMapper;
+
+	@Autowired
+	private ContextDescriptionServiceImpl contextService;
 
 	@Override
 	public void initialize() throws InitializationException {
@@ -50,11 +58,11 @@ public class InitializationServiceImpl implements InitializationService {
 			throws InitializationException {
 		LOGGER.info("Start update process definition ...");
 		if (documentContent == null || StringUtils.isEmpty(deploymentName)) {
-			throw new IllegalArgumentException("Name and document required");
+			throw new IllegalArgumentException(ErrorMessageConstants.NULL_OBJECT);
 		}
 		if (!documentContent.getContentType().equalsIgnoreCase(MimeTypeUtils.APPLICATION_XML_VALUE)
 				&& !documentContent.getContentType().equalsIgnoreCase(MimeTypeUtils.TEXT_XML_VALUE)) {
-			throw new IllegalArgumentException("Invalide type mime(" + documentContent.getContentType() + ")");
+			throw new IllegalArgumentException(ErrorMessageConstants.ILLEGAL_ATTRIBUTE);
 		}
 		RepositoryService repositoryService = processEngine.getRepositoryService();
 		if (documentContent.isFile()) {
@@ -84,39 +92,68 @@ public class InitializationServiceImpl implements InitializationService {
 
 	@Override
 	public List<org.georchestra.signalement.core.dto.ProcessDefinition> searchProcessDefinitions() {
-		List<org.georchestra.signalement.core.dto.ProcessDefinition> result = null;
 		RepositoryService repositoryService = processEngine.getRepositoryService();
 		List<ProcessDefinition> processDefinitions = repositoryService.createProcessDefinitionQuery().list();
-		if (CollectionUtils.isNotEmpty(processDefinitions)) {
-			result = new ArrayList<>(processDefinitions.size());
-			for (ProcessDefinition processDefinition : processDefinitions) {
-				result.add(processDefinitionMapper.entityToDto(processDefinition));
-			}
-		}
-		return result;
+		return processDefinitions.stream().map(processDefinitionMapper::entityToDto).collect(Collectors.toList());
 	}
 
 	@Override
-	public boolean deleteProcessDefinition(String processDefinitionName, Integer version) throws InitializationException {
-		boolean result = false;
+	public boolean deleteProcessDefinition(String processDefinitionName, Integer version) throws InvalidDataException {
+		boolean result = true;
 		RepositoryService repositoryService = processEngine.getRepositoryService();
 		ProcessDefinitionQuery query = repositoryService.createProcessDefinitionQuery()
 				.processDefinitionName(processDefinitionName);
-		if(version != null){
-			LOGGER.info("A version has been provided : ", version);
+		if (version != null) {
+			LOGGER.info("A version has been provided : {}", version);
 			query = query.processDefinitionVersion(version);
 		}
 		List<ProcessDefinition> processDefinitions = query.list();
 		if (CollectionUtils.isNotEmpty(processDefinitions)) {
+			TaskService taskService = processEngine.getTaskService();
+			List<Task> tasks = taskService.createTaskQuery().list();
+			List<ContextDescription> contexts = contextService.searchContextDescriptions(null, null);
+			ProcessDefinition latestVersion = getMostRecentVersion(processDefinitions);
 			for (ProcessDefinition processDefinition : processDefinitions) {
-				LOGGER.info("Start delete deployment {}.", processDefinition.getDeploymentId());
-				repositoryService.deleteDeployment(processDefinition.getDeploymentId(), true);
-				LOGGER.info("Delate deployment {} done.", processDefinition.getDeploymentId());
+				if (!processDefinitionIsUsed(processDefinition, latestVersion, tasks, contexts)) {
+					LOGGER.info("Start delete deployment {}.", processDefinition.getDeploymentId());
+					repositoryService.deleteDeployment(processDefinition.getDeploymentId(), true);
+					LOGGER.info("Delate deployment {} done.", processDefinition.getDeploymentId());
+				} else {
+					LOGGER.info("Can't delete used workflow : {}.", processDefinition.getDeploymentId());
+					result = false;
+				}
 			}
-			result = true;
+		} else {
+			throw new IllegalArgumentException(ErrorMessageConstants.NULL_OBJECT);
 		}
 		return result;
 	}
 
+	private ProcessDefinition getMostRecentVersion(List<ProcessDefinition> processDefinitions) {
+		return processDefinitions.stream()
+				.sorted(Comparator.comparingInt(ProcessDefinition::getVersion).reversed())
+				.collect(Collectors.toList())
+				.get(0);
+	}
 
+	private boolean processDefinitionIsUsed(ProcessDefinition processDefinition,
+											ProcessDefinition latestVersion,
+											List<Task> tasks,
+											List<ContextDescription> contexts) {
+		boolean usedInTask = tasks.stream().anyMatch(task ->
+				task.getProcessDefinitionId().equals(processDefinition.getId()));
+
+		boolean lastVersion = latestVersion.equals(processDefinition);
+		boolean usedInContext = contexts.stream().anyMatch(context ->
+				contextUsesProcess(context, processDefinition, lastVersion));
+
+		return usedInTask || usedInContext;
+	}
+
+	private boolean contextUsesProcess(ContextDescription context, ProcessDefinition process, boolean lastVersion) {
+		boolean key = context.getProcessDefinitionKey().equals(process.getKey());
+		boolean revision = (context.getRevision() != null && context.getRevision() == process.getVersion())
+				|| (context.getRevision() == null && lastVersion);
+		return key && revision;
+	}
 }
